+++ lib/honeybadger/backend/debug.rb
@@ -8,6 +8,7 @@ module Honeybadger
     class Debug < Null
       def notify(feature, payload)
         logger.unknown("notifying debug backend of feature=#{feature}\n\t#{payload.to_json}")
+        return Response.new(ENV['DEBUG_BACKEND_STATUS'].to_i, nil) if ENV['DEBUG_BACKEND_STATUS']
         super
       end
     end