+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+* Exit with 1 from deploy command when request fails but ignore the failures in
+  the capistrano task.
+
+  *Joshua Wood*
+
 * Support Resque natively.
 
   *Joshua Wood*